Refactor authentication flow and remove login/register pages
- Removed Login and Register components, streamlining the authentication process to use MediaWiki OAuth directly.
- Updated Home component to handle user authentication and redirect accordingly.
- Adjusted router logic to redirect to OAuth login instead of login/register pages.
- Enhanced OAuth callback handling to redirect users to the home page on authentication failure.
